+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `\Tools\TestSuite\TestTrait::assertException()` new performs a strict comparison and does not
   consider parent classes (`ErrorException` != `Exception`). It correctly ignores the  deprecations
   in determining the exception;
+* all methods of `\Tools\Exceptionist` no longer accept a value of `null` as `$message` parameter;
 * added many `@template` tags to the `Exceptionist`, that now always refers to `Exception` and no
   longer to `Throwable`;
 * improved the `array_clean()` global function.